.dagger/pyproject.toml
[project]
name = "operator-flows"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = ["dagger-io ==0.18.4"]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
